Refactoring pylossless.py to reduce file size.
